force runbackend to run as non-root

+++ wopl_test/bin/runbackend.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env bash
 
+if [[ $(whoami) != $WOPL_TEST_USER ]]; then
+  echo "Should run as $WOPL_TEST_USER"
+  exit 2
+fi
+
 cd $WOPL_TEST_HOME
 
 rm -rf venv
